fix/ get all tips
